+++ dgisim/src/character/characters_set.py
@@ -16,6 +16,7 @@ _DEFAULT_CHARACTERS: list[type[Character]] = [
     Keqing,
     Kaeya,
     Klee,
+    Maguukenki,
     Mona,
     Nahida,
     Noelle,