+++ dgisim/tests/test_characters/test_sangonomiya_kokomi.py
@@ -25,7 +25,7 @@ class TestSangonomiyaKokomi(unittest.TestCase):
         self.assertIn(Element.HYDRO, p2ac.get_elemental_aura())
 
     def test_elemental_skill1(self):
-        # test elemental skill generate status without damage or energy recharge
+        # test elemental skill generate summon and applies hydro aura
         game_state = step_skill(
             self.BASE_GAME,
             Pid.P1,